pyproject.toml
[tool.poetry]
name = "NDE"
version = "0.1.0"
description = "Neural differential equations usage"
authors = ["Golov V. A."]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.12"
numpy = "^1.23.0"
matplotlib = "^3.5.2"
scipy = "^1.8.1"
seaborn = "^0.11.2"
pytorch = "^1.0.2"
torchdyn = "^1.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
